# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "e-commerce site written by go",
        "title": "Go Shopping Gin API",
        "termsOfService": "https://github.com/xguox/coconut",
        "contact": {
            "name": "XguoX",
            "url": "https://xguox.me"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9876",
    "basePath": "/api/v1",
    "paths": {
        "/admin/categories": {
            "post": {
                "description": "添加新的商品分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "添加新的商品分类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证 Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建分类请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.CategoryValidator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{msg:\"请求处理成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "{msg:\"请求参数有误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{msg:\"服务器错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "description": "后台账号登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "后台账号登录",
                "parameters": [
                    {
                        "description": "账号登录请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.LoginValidator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{msg:\"请求处理成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{msg:\"账号或密码有误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "{msg:\"请求参数有误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{msg:\"服务器错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CategoryValidator": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "object",
                    "required": [
                        "name",
                        "slug"
                    ],
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "slug": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "model.LoginValidator": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "required": [
                        "email",
                        "password"
                    ],
                    "properties": {
                        "email": {
                            "type": "string"
                        },
                        "password": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}